+++ src/bridge/fly/cache.ts
@@ -4,8 +4,10 @@ import log from '../../log'
 import { transferInto } from '../../utils/buffer'
 import { Bridge } from '../bridge';
 import { Runtime } from '../../runtime';
+import { CacheOperation, isCacheOperation } from '../../cache_notifier';
 
 const errCacheStoreUndefined = new Error("cacheStore is not defined in the config.")
+const errCacheNotifierUndefined = new Error("cacheNotifier is not defined in the config")
 
 registerBridge('flyCacheSet', function cacheSet(rt: Runtime, bridge: Bridge, key: string, value: ArrayBuffer | string, options: string | undefined, callback: ivm.Reference<Function>) {
   if (!bridge.cacheStore) {
@@ -27,7 +29,7 @@ registerBridge('flyCacheSet', function cacheSet(rt: Runtime, bridge: Bridge, key
     callback.applyIgnored(null, [err.toString()])
     return
   }
-  bridge.cacheStore.set(rt, key, buf, opts).then((ok) => {
+  bridge.cacheStore.set(rt.app.id, key, buf, opts).then((ok) => {
     rt.reportUsage("cache:set", { size: buf.byteLength })
     callback.applyIgnored(null, [null, ok])
   }).catch((err) => {
@@ -43,7 +45,7 @@ registerBridge('flyCacheExpire', function cacheExpire(rt: Runtime, bridge: Bridg
     return
   }
 
-  bridge.cacheStore.expire(rt, key, ttl).then((ok) => {
+  bridge.cacheStore.expire(rt.app.id, key, ttl).then((ok) => {
     callback.applyIgnored(null, [null, ok])
   }).catch((err) => {
     callback.applyIgnored(null, [err.toString()])
@@ -57,7 +59,7 @@ registerBridge('flyCacheGet',
       return
     }
 
-    bridge.cacheStore.get(rt, key).then((buf) => {
+    bridge.cacheStore.get(rt.app.id, key).then((buf) => {
       rt.reportUsage("cache:get", { size: buf ? buf.byteLength : 0 })
       callback.applyIgnored(null, [null, transferInto(buf)])
     }).catch((err) => {
@@ -73,11 +75,11 @@ registerBridge('flyCacheDel',
       return
     }
 
-    bridge.cacheStore.del(rt, key).then((result) => {
+    bridge.cacheStore.del(rt.app.id, key).then((result) => {
       callback.applyIgnored(null, [null, !!result])
     }).catch((err) => {
       log.error("got err in cache.del", err)
-      callback.applyIgnored(null, [null, null]) // swallow errors on get for now
+      callback.applyIgnored(null, [err.toString()])
     })
   })
 
@@ -88,11 +90,11 @@ registerBridge('flyCacheSetTags',
       return
     }
 
-    bridge.cacheStore.setTags(rt, key, tags).then((result) => {
+    bridge.cacheStore.setTags(rt.app.id, key, tags).then((result) => {
       callback.applyIgnored(null, [null, result])
     }).catch((err) => {
       log.error("got err in cache.setTags", err)
-      callback.applyIgnored(null, [null, null]) // swallow errors on get for now
+      callback.applyIgnored(null, [err.toString()])
     })
   })
 
@@ -103,12 +105,32 @@ registerBridge('flyCachePurgeTags',
       return
     }
 
-    bridge.cacheStore.purgeTags(rt, key).then((result) => {
+    bridge.cacheStore.purgeTag(rt.app.id, key).then((result) => {
       setImmediate(() => {
         callback.applyIgnored(null, [null, JSON.stringify(result)])
       })
     }).catch((err) => {
       log.error("got err in cache.purgeTags", err)
-      callback.applyIgnored(null, [null, null]) // swallow errors on get for now
+      callback.applyIgnored(null, [err.toString()])
+    })
+  })
+
+
+registerBridge('flyCacheNotify',
+  function cacheDel(rt: Runtime, bridge: Bridge, type: string | CacheOperation, key: string, callback: ivm.Reference<Function>) {
+    if (!bridge.cacheStore || !bridge.cacheNotifier) {
+      callback.applyIgnored(null, [errCacheNotifierUndefined.toString()])
+      return
+    }
+
+    if (!isCacheOperation(type)) {
+      callback.applyIgnored(null, ["Invalid cache notification type"])
+      return
+    }
+    bridge.cacheNotifier.send(type, rt.app.id, key).then((result) => {
+      callback.applyIgnored(null, [null, !!result])
+    }).catch((err) => {
+      log.error("got err in cacheNotify", err)
+      callback.applyIgnored(null, [err.toString()])
     })
   })