+++ test/js-fails.spec.ts
@@ -22,7 +22,7 @@ describe('JS Fails', function () {
     describe("setTimeout fires after response", function () {
       before(startServer(`fails/async-app`))
       before(function () {
-        return cacheStore.set(this.server.runtime, "long-wait-after-response", "no")
+        return cacheStore.set(this.server.runtime.app.id, "long-wait-after-response", "no")
       })
       after(stopServer)
 
@@ -44,7 +44,7 @@ describe('JS Fails', function () {
 
         await (sleep(150))
 
-        let cached = await cacheStore.get(this.server.runtime, "long-wait-after-response")
+        let cached = await cacheStore.get(this.server.runtime.app.id, "long-wait-after-response")
         expect((<Buffer>cached).toString()).to.equal(cacheValue)
       })
     })