+++ v8env/src/fly/cache/index.ts
@@ -19,7 +19,8 @@
 declare var bridge: any
 export interface CacheSetOptions {
   ttl?: number,
-  tags?: string[]
+  tags?: string[],
+  onlyIfEmpty?: boolean
 }
 
 
@@ -147,8 +148,26 @@ export function purgeTag(tag: string) {
  * @returns true if delete was successful
  */
 export function del(key: string) {
-  return expire(key, 0)
+  return new Promise<boolean>(function cacheDelPromise(resolve, reject) {
+    bridge.dispatch("flyCacheDel", key, function cacheDelCallback(err: string | null, ok?: boolean) {
+      if (err != null) {
+        reject(err)
+        return
+      }
+      resolve(ok)
+    })
+  })
 }
+
+/**
+ * A library for caching/retrieving Response objects
+ */
+export { default as responseCache } from "./response"
+
+/**
+ * Global cache operations
+ */
+import { default as global } from "./global"
 
 const cache = {
   get,
@@ -157,11 +176,7 @@ const cache = {
   expire,
   del,
   setTags,
-  purgeTag
+  purgeTag,
+  global
 }
 export default cache
-
-/**
- * A library for caching/retrieving Response objects
- */
-export { default as responseCache } from "./response"